backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "LocalWriter Backend"
authors = [
    {name = "unser229", email = "nzlobin041@gmail.com"}
]
readme = "README.md"
# Разрешаем Python от 3.11 до 3.13 (исключая 3.13, так как там могут быть проблемы с колесами)
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-multipart",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "chromadb (>=1.4.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
# Эта секция дублирует ограничения для старых версий poetry, пусть совпадает
python = ">=3.11,<3.13"